+++ src/plone/app/blob/tests/base.py
@@ -1,14 +1,10 @@
 import unittest
 from plone.app.blob.tests.layer import BlobLayer, BlobReplacementLayer
 from plone.app.blob.tests.layer import BlobLinguaLayer
-try:
-    # try to import the sample type for testing LinguaPlone
-    from plone.app.blob.tests import lingua
-    lingua      # make pyflakes happy...
-except ImportError:
-    pass
+from plone.app.testing import TEST_USER_NAME
+from plone.app.testing import TEST_USER_PASSWORD
 
-
+from plone.testing.z2 import Browser
 
 
 class BlobTestCase(unittest.TestCase):
@@ -27,6 +23,17 @@ class ReplacementTestCase(unittest.TestCase):
     """ base class for integration tests using replacement types """
 
     layer = BlobReplacementLayer
+
+    def getCredentials(self):
+        return '%s:%s' % (TEST_USER_NAME, TEST_USER_PASSWORD)
+
+    def getBrowser(self, loggedIn=True):
+        """ instantiate and return a testbrowser for convenience """
+        browser = Browser(self.layer['app'])
+        if loggedIn:
+            auth = 'Basic %s' % self.getCredentials()
+            browser.addHeader('Authorization', auth)
+        return browser
 
 
 class ReplacementFunctionalTestCase(unittest.TestCase):